+++ server/app.js
@@ -1,8 +1,8 @@
 import 'dotenv/config';
 import express from "express";
 import bodyParser from "body-parser";
-import router from './routes/user';
-import messageRouter from './routes/message';
+import userRouter from './routes/users';
+import messageRouter from './routes/messages';
 
 
 const PORT = 3000;
@@ -12,7 +12,7 @@ const app = express();
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
 
-app.use('/api/v1', router);
+app.use('/api/v1/auth', userRouter);
 app.use('/api/v1/messages', messageRouter);
 
 app.get('/', (req, res) => {
